ga1/q18/semantic-search/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.54.3
numpy==1.26.3
python-dotenv==1.0.0
scikit-learn==1.4.0
requests==2.31.0
httpx==0.27.0